updating the test function to do not have to create a server listening in a go routine
